Lutris Doc. Add note about docker named volumes.
Added a section below the first config.toml example noting that
the sugested configuration creates a docker named-volume, and
that the user has the option to use an arbitrary system directory
intead.

Also added a new config.toml to the end to tell the user to remove
a now redundent environment variable.