ci(docker-build-and-publish): improve caching logic
We try if the current repo (which might be a fork) has a build-cache image, if not, we fall back to the games-on-whales one.
For PRs we prioritize the local cache, then the repo cache and lastly the games-on-whales cache.
This should speed up the first time build vastly as we can fall back to at least one existing build cache.