ci(docker-build-and-publish): ignore conclusion
We are already looking for a commit so our chance of finding the correct artifact should be high enough to ignore if the workflow is in_progress or not.
My guess is that if one job fails early there is a chance that the workflow_conclusion is failed which is why in_progress artifacts are not found anymore.