ci(docker-build-and-publish): enable workflows for pull requests (#237)
* ci(build-wildlife): restructure event dispatch logic to check for GHCR_TOKEN

* ci(docker-build-and-publish): use default token

GitHub Workflows already populate a GITHUB_TOKEN for pushing to the registry. Using github.actor this should allow all contributers with permission to build and upload.

* ci(docker-build-and-publish): allow running arm64 build with x86 runner

* ci(docker-build-and-publish): allow merge without tokens

* ci(auto-build): allow run on PRs

* ci(docker-build-and-publish): fix redunant ghcr.io definition for registry image

* ci(docker-build-and-publish): allow transferring builds for pull requests

Using a combination of GitHub artifacts and a local registry we can
transfer built images between all jobs in the workflow.
This allows building the image in pull requests to verify that the build
succeeds before merging the pull request.
Sadly this is the only way to do this as GitHub does not allow the
images to be pushed to the fork maintainers GHCR which would make this
much simpler.

* ci(docker-build-and-publish): improve caching logic

We try if the current repo (which might be a fork) has a build-cache image, if not, we fall back to the games-on-whales one.
For PRs we prioritize the local cache, then the repo cache and lastly the games-on-whales cache.
This should speed up the first time build vastly as we can fall back to at least one existing build cache.

* ci(docker-build-and-publish): ignore conclusion

We are already looking for a commit so our chance of finding the correct artifact should be high enough to ignore if the workflow is in_progress or not.
My guess is that if one job fails early there is a chance that the workflow_conclusion is failed which is why in_progress artifacts are not found anymore.